Modularização e divisão entre HPP e CPP
