+++ Empregado.hpp
@@ -1,7 +1,6 @@
 #ifndef EMPREGADO_H
 #define EMPREGADO_H
 
-#include <iostream>
 #include <string>
 
 class Empregado {
@@ -9,18 +8,9 @@ class Empregado {
 	public:
 		double salarioHora;
 		double quotaMensalVendas;
+		std::string nome;
 
-	double pagamentoMes(double horasTrabalhadas) {
- 
-	  double t = horasTrabalhadas;
-	  
-	  //Cálculo de hora extra (+50% se horasTrabalhadas > 8)
-	  if (horasTrabalhadas > 8) {
-		double x = horasTrabalhadas - 8;
-		t += x / 2;
-	  }
-	  return t * salarioHora;
-	}
+		double pagamentoMes(double horasTrabalhadas);
 	
 };
