--- Engenheiro.cpp
@@ -1,11 +1,0 @@
-#include <string>
-#include "Empregado.hpp"
-
-class Engenheiro : public Empregado {
-
-	public:
-		std::string nome;  
-	int projetos;
-	
-};
-
