+++ Engenheiro.hpp
@@ -1,0 +1,9 @@
+#include "Empregado.hpp"
+
+class Engenheiro : public Empregado {
+
+	public:
+
+	int projetos;
+
+};
