+++ Vendedor.cpp
@@ -1,13 +1,5 @@
-#include <string>
-#include "Empregado.hpp"
+#include "Vendedor.hpp"
 
-class Vendedor : public Empregado {
-
-	public:
-		std::string nome;  	  
-
-		double quotaTotalAnual() {
-		return quotaMensalVendas * 12;
-		}
-	
-};
+double Vendedor::quotaTotalAnual() {
+	return quotaMensalVendas * 12;
+}
