+++ Vendedor.hpp
@@ -1,0 +1,9 @@
+#include "Empregado.hpp"
+
+class Vendedor : public Empregado {
+
+	public:
+
+		double quotaTotalAnual();
+
+};
